Babel config, added packages, fixed submission model
Added config for Babel in an attempt to not have it compile the
node_modules folder. Added the packages that are dependencies for the
tools.
The submission model change is unfortunate. This was introduced in the
rewrite with the tools which makes the rewritten services somewhat
incompatible with the original services.