Rewrite without ForkingStore, removed more packages
When removing unused packages, the ForkingStore jumped out as only being
used sparingly and something that does not provide much documentation.
So this was removed in favour of some basic transformations with N3
stores.
I'm sorry, but if there is no documentation on a library, I'm not using
it, and I will convert code to using libraries that are properly
documented and well tested. I don't want to debug something like that
otherwise.