Add the `DecisionArticles` validator
This ensures we use the same validations as the frontend. We simply copy paste the implementation for now since it is unlikely to change much once it is stable and it allows us to do extra things in the backend version if needed.

Related code: https://github.com/lblod/frontend-loket/pull/391